requirements.txt
pydantic>=2.0.0
pytest
chromadb>=0.4.0
google-generativeai
python-dotenv
langchain-text-splitters
matplotlib
